ai-contact-center-solution-accelerator-main/ai-contact-center-solution-accelerator-main/pyproject.toml
[project]
name = "ai-contact-centre-solution-accelerator"
version = "1.0.0"
description = "Multi-agent realtime voice assistant with intelligent handoffs"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "fastapi[standard]>=0.116.1",
    "semantic-kernel[mcp,realtime]>=1.35.3",
    "azure-identity>=1.23.0",
    "azure-communication-callautomation>=1.4.0",
    "azure-communication-email>=1.0.0",
    "fastmcp>=2.0.0",
    "azure-eventgrid>=4.20.0",
    "pydantic>=2.0.0,<3.0.0",
    "pyyaml>=6.0.0,<7.0.0",
    "numpy>=1.26.0",
    "sounddevice>=0.5.0,<1.0.0",
    "PyJWT>=2.8.0,<3.0.0",
    "cryptography>=42.0.0,<47.0.0",
    "pydantic-settings>=2.0.0,<3.0.0",
]

[build-system]
requires = ["uv_build>=0.9.20,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=6.0.0",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.9.3",
    "ty>=0.0.12",
    "pre-commit>=4.0.0",
]
scripts = [
    "azure-communication-phonenumbers>=1.2.0",
    "azure-identity>=1.23.0",
]

[tool.ruff]
src = ["src"]
line-length = 120

[tool.ruff.lint]
extend-select = ["I"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
python_classes = ["Test*"]
filterwarnings = [
    "ignore::pydantic.PydanticDeprecatedSince211",
]

[tool.coverage.run]
omit = ["tests/**"]

[tool.ty.src]
include = ["src"]

[tool.uv.workspace]
members = [".", "eval", "examples/mcp-servers/customer-verification"]


